@@ -1,9 +1,8 @@
 using System.Net;
 using System.Net.Sockets;
 
-// You can use print statements as follows for debugging, they'll be visible when running tests.
 Console.WriteLine("Starting server...");
 
-TcpListener server = new TcpListener(IPAddress.Any, 4221);
+TcpListener server = new(IPAddress.Any, 4221);
 server.Start();
-server.AcceptSocket(); // wait for client
+server.AcceptSocket(); // wait for a client
